+++ packages/openapi-ts/src/index.ts
@@ -11,21 +11,26 @@ import type { Config, UserConfig } from './types/config';
 import { registerHandlebarTemplates } from './utils/handlebars';
 import { Performance, PerformanceReport } from './utils/performance';
 
+type Configs = UserConfig | (() => UserConfig) | (() => Promise<UserConfig>);
+
 /**
  * Generate a client from the provided configuration.
  *
  * @param userConfig User provided {@link UserConfig} configuration.
  */
 export const createClient = async (
-  userConfig: UserConfig,
+  userConfig?: Configs,
 ): Promise<ReadonlyArray<Client | IR.Context>> => {
+  const resolvedConfig =
+    typeof userConfig === 'function' ? await userConfig() : userConfig;
+
   let configs: Config[] = [];
 
   try {
     Performance.start('createClient');
 
     Performance.start('config');
-    configs = await initConfigs(userConfig);
+    configs = await initConfigs(resolvedConfig);
     Performance.end('config');
 
     Performance.start('handlebars');
@@ -61,11 +66,12 @@ export const createClient = async (
     return result;
   } catch (error) {
     const config = configs[0] as Config | undefined;
-    const dryRun = config ? config.dryRun : userConfig?.dryRun;
+    const dryRun = config ? config.dryRun : resolvedConfig?.dryRun;
+
     // TODO: add setting for log output
     if (!dryRun) {
-      const logs = config?.logs ?? getLogs(userConfig);
-      if (logs.level !== 'silent') {
+      const logs = config?.logs ?? getLogs(resolvedConfig);
+      if (logs.level !== 'silent' && logs.file) {
         const logName = `openapi-ts-error-${Date.now()}.log`;
         const logsDir = path.resolve(process.cwd(), logs.path ?? '');
         ensureDirSync(logsDir);
@@ -82,11 +88,15 @@ export const createClient = async (
 /**
  * Type helper for openapi-ts.config.ts, returns {@link UserConfig} object
  */
-export const defineConfig = (config: UserConfig): UserConfig => config;
+export const defineConfig = async (config: Configs): Promise<UserConfig> =>
+  typeof config === 'function' ? await config() : config;
 
 export { defaultPlugins } from './initConfigs';
 export type { IR } from './ir/types';
 export type { OpenApi } from './openApi/types';
+export { clientDefaultConfig } from './plugins/@hey-api/client-core/config';
+export { clientPluginHandler } from './plugins/@hey-api/client-core/plugin';
+export type { Client } from './plugins/@hey-api/client-core/types';
 export type { Plugin } from './plugins/types';
 export type { UserConfig } from './types/config';
 export type { LegacyIR } from './types/types';