+++ examples/openapi-ts-nuxt/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @example/openapi-ts-nuxt
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hey-api/nuxt@0.1.4
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hey-api/nuxt@0.1.3
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`0670a38`](https://github.com/hey-api/openapi-ts/commit/0670a38c66dffc1431a7d4b48ec06b72d4f6868f)]:
+  - @hey-api/nuxt@0.1.2
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @hey-api/nuxt@0.1.1
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`d80f835`](https://github.com/hey-api/openapi-ts/commit/d80f835b46775a01451f02f832ceb288c2b561d2)]:
+  - @hey-api/nuxt@0.1.0
+
 ## 0.0.10
 
 ### Patch Changes